Initial file separation.
Pending add tests to ensure it is still working right
